Disable docker-compoes's tty allocation
Add the -T flag to `docker-compose`  which disables its tty allocation [1] (the inverse of docker's -t flag). This prevents the staircase formatting we were observing when printing (not just logging) messages received by the master controler's Listner (which is running in a background thread). Because this container runs in the forground, it can change the tty's settings. This is a problem because it looks like the container disables opost (which postprocess output) in the tty, which screws up the formatting done with onlcr [2].

Also, we don't need a tty because logging to stdout/stderr directly in these processes is not threadsafe.

A simple example of this error can be seen with the 'Sending and receiving logging events across a network' example in the python logging cookbook, which sets up a handler on the server to listen for messages and a handler on the client to send them. Start the server with

$ python3 server.py &

and the client with:

$ docker run -i --rm -v /home/sgrichar/terra/terra_dsm/external/terra:/src python:3 python3 /src/client.py

which looks as expected; compared to this, which doesn't:

$ docker run -it --rm -v /home/sgrichar/terra/terra_dsm/external/terra:/src python:3 python3 /src/client.py

[1] https://github.com/moby/moby/issues/8513
[2] https://unix.stackexchange.com/a/242814/332869
[3] https://docs.python.org/3/howto/logging-cookbook.html#sending-and-receiving-logging-events-across-a-network